Move get_pauli_basis
