+++ test/unit/decoders/executor_estimator/test_utils.py
@@ -17,6 +17,7 @@ from qiskit.quantum_info import Pauli
 
 from qiskit_ibm_runtime.decoders.executor_estimator.utils import (
     compute_exp_val,
+    get_pauli_basis,
     identify_measure_basis,
     project_to_z,
 )
@@ -223,3 +224,28 @@ class TestComputeExpVal(IBMTestCase):
                     twirl_var,
                     decimal=10,
                 )
+
+
+class TestGetPauliBasis(IBMTestCase):
+    """Tests for get_pauli_basis function."""
+
+    def test_single_qubit_bases(self):
+        """Test single-qubit basis conversions."""
+        for basis, expected in [
+            ("0", Pauli("Z")),
+            ("1", Pauli("Z")),
+            ("+", Pauli("X")),
+            ("-", Pauli("X")),
+            ("r", Pauli("Y")),
+            ("l", Pauli("Y")),
+            ("I", Pauli("I")),
+        ]:
+            with self.subTest(basis=basis):
+                result = get_pauli_basis(basis)
+                self.assertEqual(result, expected)
+
+    def test_multi_qubit(self):
+        """Test multi-qubit basis conversion."""
+        result = get_pauli_basis("0+r")
+        expected = Pauli("ZXY")
+        self.assertEqual(result, expected)
